refactor: rename all files to kebab case
