--- src/components/Dialog/EditDialog/EditDialog.test.tsx
@@ -1,25 +1,0 @@
-import { render, RenderResult } from '@testing-library/react'
-import { afterAll, beforeAll, describe, expect, test } from 'vitest'
-
-import CommonEditDialog from '@/components/Dialog/EditDialog/EditDialog'
-
-test('CommonEditDialog', () => {
-  describe('CommonEditDialog', () => {
-    let DialogMounted: RenderResult
-
-    beforeAll(() => {
-      DialogMounted = render(
-        <CommonEditDialog
-          itemToEdit={{ id: '' }}
-          editHandler={(input) => {
-            expect(input)
-          }}
-        />
-      )
-    })
-
-    afterAll(() => {
-      DialogMounted.unmount()
-    })
-  })
-})
