+++ src/components/retro-view/views/discussing-phase/components/item-stack.tsx
