+++ src/components/retro-view/views/voting-phase/components/item-voter.tsx
