+++ src/server/trpc/router/index.ts
@@ -1,6 +1,6 @@
 // src/server/trpc/router/index.ts
-import { retroItemRouter } from '@/server/trpc/router/retroItem'
-import { retroItemCollectionRouter } from '@/server/trpc/router/retroItemCollection'
+import { retroItemRouter } from '@/server/trpc/router/retro-item'
+import { retroItemCollectionRouter } from '@/server/trpc/router/retro-item-collection'
 import { retrospectiveRouter } from '@/server/trpc/router/retrospective'
 import { userRouter } from '@/server/trpc/router/user'
